+++ src/JlWrap/io.jl
@@ -102,7 +102,7 @@ function pyjlbinaryio_readinto(io::IO, b::Py)
         return PyNULL
     end
     pydel!(c)
-    buf = unsafe_load(C.PyMemoryView_GET_BUFFER(getptr(m)))
+    buf = unsafe_load(C.PyMemoryView_GET_BUFFER(m))
     if buf.readonly != 0
         pydel!(m)
         errset(pybuiltins.ValueError, "output buffer is read-only")
@@ -125,7 +125,7 @@ function pyjlbinaryio_write(io::IO, b::Py)
         return PyNULL
     end
     pydel!(c)
-    buf = unsafe_load(C.PyMemoryView_GET_BUFFER(getptr(m)))
+    buf = unsafe_load(C.PyMemoryView_GET_BUFFER(m))
     data = unsafe_wrap(Array, Ptr{UInt8}(buf.buf), buf.len)
     write(io, data)
     pydel!(m)
@@ -337,7 +337,6 @@ Wrap `io` as a Python binary IO object.
 This is the default behaviour of `Py(io)`.
 """
 pybinaryio(v::IO) = pyjl(pyjlbinaryiotype, v)
-export pybinaryio
 
 """
     pytextio(io::IO)
@@ -345,6 +344,5 @@ export pybinaryio
 Wrap `io` as a Python text IO object.
 """
 pytextio(v::IO) = pyjl(pyjltextiotype, v)
-export pytextio
 
 pyjltype(::IO) = pyjlbinaryiotype